refactor : create LineStationService class and LineStationController class
